+++ packages/form/types/LmCheckbox.d.ts
@@ -1,8 +1,9 @@
 import { ReactNode } from 'react';
-import { ParagraphProps, SizeTokens, XStackProps } from 'tamagui';
+import { CheckboxProps, CheckedState, ParagraphProps, SizeTokens, XStackProps } from 'tamagui';
 export type LmCheckboxProps = XStackProps & {
+    checkboxProps?: CheckboxProps;
     label?: ReactNode;
-    onChange?: (bool: boolean) => void;
+    onChange?: (state: CheckedState) => void;
     value?: boolean;
     error?: boolean;
     helperText?: ReactNode;
@@ -10,5 +11,5 @@ export type LmCheckboxProps = XStackProps & {
     size?: SizeTokens;
     required?: boolean;
 };
-export declare function LmCheckbox({ label, onChange, error, helperText, size, value, required, theme, helperTextProps, ...stackProps }: LmCheckboxProps): JSX.Element;
+export declare function LmCheckbox({ label, onChange, error, helperText, size, value, required, theme, helperTextProps, checkboxProps, ...stackProps }: LmCheckboxProps): JSX.Element;
 //# sourceMappingURL=LmCheckbox.d.ts.map